Adding react and typescript basic setup
To start working on the application, I added both our frontend engine (React & ReactDOM) and Typescript, so we can work with type checking over JS. Also, I added React Hot Loader, which should make the application loading times faster (I am concerned with the fact the library uses NodeModule, I hope the lib has a @types definition that solves the usage of the default module; it is not ideal to cast to any, since I'm using TS... duh), and added the Parcel Bundler (this is a small project, I see no need to get involved with resolved Webpack configurations as of now). Last, but not least, I added a test message on the app's main file and made a minor change to the reference code from Spotify.